requirements.txt
# local package
-e .
pillow
dvc
"dvc[gdrive]"
torch
torchvision
matplotlib
hydra-core
wandb
omegaconf
pytorch_lightning

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
